chat_bot/requirements.txt
aiohttp==3.9.3
channels==4.0.0
daphne==4.1.0
Django==4.2.11
djangorestframework==3.15.0
environs==11.0.0
pyTelegramBotAPI==4.16.1
pytest==8.1.1
pytest-django==4.8.0
pytest-lazy-fixtures==1.0.7
python-json-logger==2.0.7
psycopg2-binary==2.9.9
redis==5.0.3
django-cors-headers==4.3.1
drf-yasg==1.21.7